+++ ennemies.ino
@@ -2,7 +2,7 @@
 
 boolean collideEnnemyAndBullets(Ennemies ennemy) 
 {
-  for(uint8_t i = 0 ; i< NB_PLAYER_BULLET; ++i){
+  for(int i = 0 ; i< NB_PLAYER_BULLET; ++i){
     if(soldatBullet[i].life>0) {
       if(Collision(soldatBullet[i].x, soldatBullet[i].y, 2, 2, ennemy.x, ennemy.y, 5, 5)) {
         soldatBullet[i].life = 0;
@@ -16,7 +16,7 @@ boolean collideEnnemyAndBullets(Ennemies ennemy)
 void initEnnemies()
 {
    //init state
-  for(uint8_t i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
+  for(int i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
     ennemies[i].life = 3;
     ennemies[i].x = gb.display.width() + random(0,10); //first time we add ennemei randomly
     ennemies[i].y = random(0,gb.display.height());
@@ -25,7 +25,7 @@ void initEnnemies()
 
 void updateEnnemies()
 {
-  for(uint8_t i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
+  for(int i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
    if(ennemies[i].life>0) {
     if(collideEnnemyAndBullets(ennemies[i])) {
       --ennemies[i].life; //we die
@@ -38,7 +38,7 @@ void updateEnnemies()
 void drawEnnemies()
 {
   gb.display.setColor(RED);
-  for(uint8_t i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
+  for(int i = 0 ; i< NB_MAX_ENNEMIES; ++i) {
    if(ennemies[i].life>0) {
     gb.display.fillRect( ennemies[i].x , ennemies[i].y , 5 , 5);
    }
