+++ playerBullet.ino
@@ -1,8 +1,8 @@
 #define VELOCITY_SOLDAT_BULLET 3
 
-int16_t getNextBullet()
+int getNextBullet()
 {
-  for(int16_t i = 0 ; i< NB_PLAYER_BULLET; ++i) {
+  for(int i = 0 ; i< NB_PLAYER_BULLET; ++i) {
     if(soldatBullet[i].life == 0) {
       return i;
     }
@@ -12,7 +12,7 @@ int16_t getNextBullet()
 
 void createNewSoldatBullet()
 {
-  int16_t index = getNextBullet();
+  int index = getNextBullet();
 
   if(index != -1) {
     soldatBullet[index].life = 1;
@@ -24,7 +24,7 @@ void createNewSoldatBullet()
 void initPlayerBullet()
 {
   //init state
-  for(uint8_t i = 0 ; i< NB_PLAYER_BULLET; ++i) {
+  for(int i = 0 ; i< NB_PLAYER_BULLET; ++i) {
     soldatBullet[i].life = 0;
   }
 }
@@ -36,7 +36,7 @@ void updatePlayerBullet()
   soldat.fire = false;
    createNewSoldatBullet();
  }
-   for(uint8_t i = 0 ; i< NB_PLAYER_BULLET; ++i){
+   for(int i = 0 ; i< NB_PLAYER_BULLET; ++i){
     if(soldatBullet[i].life>0) { //if bullet is alive
       if(soldatBullet[i].x > gb.display.width()) { //if bullet is out of screen
         soldatBullet[i].life = 0;
@@ -50,7 +50,7 @@ void updatePlayerBullet()
 void drawPlayerBullet()
 {
   gb.display.setColor(YELLOW);
-   for(uint8_t i = 0 ; i< NB_PLAYER_BULLET; ++i) {
+   for(int i = 0 ; i< NB_PLAYER_BULLET; ++i) {
      if(soldatBullet[i].life>0) {
         gb.display.fillRect( soldatBullet[i].x , soldatBullet[i].y , 1 , 1);
      }
